Se arreglo el tema del archivo inexistente.

+++ src/main/java/Compilador/App.java
@@ -6,6 +6,7 @@ import Compilador.Lexico.Tipos;
 import Compilador.Sintactico.Parser;
 import Compilador.Utilidad.Logger;
 
+import javax.swing.*;
 import java.io.*;
 import java.nio.file.Files;
 import java.nio.file.Paths;
@@ -15,7 +16,7 @@ import java.util.Properties;
 
 public class App {
 
-    public static String path,filename,masm_path;
+    public static String path, filename, masm_path;
 
     public static void main(String[] args) throws IOException {
         boolean imprime = true;
@@ -27,47 +28,53 @@ public class App {
                 prop.load(in);
                 in.close();
                 masm_path = prop.getProperty("masm_path");
-            }
-            catch(Exception e) {
+            } catch (Exception e) {
                 masm_path = "C:\\masm32\\bin\\";
             }
             String data = "";
             path = file.getAbsoluteFile().getParent() + File.separator;
             filename = file.getName().split("\\.")[0];
             Logger logger = Logger.getInstance();
-            Logger.setFilename(path,filename);
+            Logger.setFilename(path, filename);
             try {
-                data = new String(Files.readAllBytes(Paths.get(file.getAbsolutePath())));
+                if (file.exists()) {
+                    data = new String(Files.readAllBytes(Paths.get(file.getAbsolutePath())));
+                    AnalizadorLexico lexico = new AnalizadorLexico(data);
+                    Parser parser = new Parser(lexico);
+                    Nodo.setLexico(lexico);
+                    Logger.setParser(parser);
+                    int ret_parser = parser.yyparse_publico();
+                    if (imprime) {
+                        logger.dumpErrors();
+                        logger.dumpEvents();
+                        logger.dumpWarnings();
+                    }
+                    System.out.println(lexico.tablaDeSimbolos.toString());
+                    if (!parser.error && ret_parser != 1) {
+                        FileWriter myFile = new FileWriter(path + filename + "-arbol-sintactico.txt");
+                        myFile.write(parser.raiz.imprimir());
+                        myFile.close();
+                        parser.raiz.generarCodigo();
+                        generarCodigo(lexico);
+                    }
+                } else {
+                    System.out.println("Archivo \"" + file.getName() + "\" inexistente.");
+                }
             } catch (Exception e) {
                 e.printStackTrace();
             }
-            AnalizadorLexico lexico = new AnalizadorLexico(data);
-            Parser parser = new Parser(lexico);
-            Nodo.setLexico(lexico);
-            Logger.setParser(parser);
-            int ret_parser=parser.yyparse_publico();
-            if (imprime) {
-                logger.dumpErrors();
-                logger.dumpEvents();
-                logger.dumpWarnings();
-            }
-            System.out.println(lexico.tablaDeSimbolos.toString());
-            if (!parser.error && ret_parser!=1) {
-                FileWriter myFile = new FileWriter(path + filename + "-arbol-sintactico.txt");
-                myFile.write(parser.raiz.imprimir());
-                myFile.close();
-                parser.raiz.generarCodigo();
-                generarCodigo(lexico);
-            }
+
         } else {
             System.out.println("Se esperaban argumentos :c ");
         }
+        System.out.println("Presione una tecla para finalizar...");
+        System.in.read();
     }
 
     public static void generarCodigo(AnalizadorLexico lexico) throws IOException {
         FileWriter myFile = new FileWriter(path + filename + ".asm");
         myFile.write(".386\n.model flat, stdcall\n.stack 200h\noption casemap :none\ninclude \\masm32\\include\\masm32rt.inc\n");
-        if(Salida.hay_salida || Llamada.hay_llamada || Division.hay_division)
+        if (Salida.hay_salida || Llamada.hay_llamada || Division.hay_division)
             myFile.write("dll_dllcrt0 PROTO C\nprintf PROTO C :VARARG");
         myFile.write("\n\n.DATA\n");
         for (Map.Entry<String, HashMap<String, Object>> entry : lexico.tablaDeSimbolos.entrySet()) {
@@ -78,8 +85,8 @@ public class App {
                     String inic = (String) entry.getValue().get("Inic");
                     myFile.write("_" + entry.getKey() + " DW " + inic + "\n");
                 } else {
-                    if(entry.getValue().get("NI")!=entry.getValue().get("Contador"))
-                    myFile.write("_" + entry.getKey().replace("-", "N") + " DW " + entry.getKey() + "\n");
+                    if (entry.getValue().get("NI") != entry.getValue().get("Contador"))
+                        myFile.write("_" + entry.getKey().replace("-", "N") + " DW " + entry.getKey() + "\n");
                 }
             } else if (tipo == Tipos.FLOAT) {
                 if (uso == "variable") {
@@ -91,26 +98,26 @@ public class App {
                     myFile.write("_" + entry.getKey().replace(".", "_").replace("-", "N") + " DQ " + entry.getKey() + "\n");
                 }
             } else if (tipo == Tipos.STRING) {
-                myFile.write("_" + entry.getKey().replaceAll("\'", "").replaceAll(" ","_") + " DB " + entry.getKey() + ",0\n");
-            }else {
-                if(uso=="procedimiento"){
-                    String nombre_proc= entry.getKey();
-                    myFile.write("NI_"+nombre_proc+ " DW " +lexico.tablaDeSimbolos.get(nombre_proc).get("NI") +"\n");
-                    myFile.write("FLAG_"+nombre_proc+ " DW 0\n");
-                    myFile.write("nombre_"+nombre_proc+ " DB " + "\'"+nombre_proc.substring(0,nombre_proc.indexOf("@")) +"\',0\n");
+                myFile.write("_" + entry.getKey().replaceAll("\'", "").replaceAll(" ", "_") + " DB " + entry.getKey() + ",0\n");
+            } else {
+                if (uso == "procedimiento") {
+                    String nombre_proc = entry.getKey();
+                    myFile.write("NI_" + nombre_proc + " DW " + lexico.tablaDeSimbolos.get(nombre_proc).get("NI") + "\n");
+                    myFile.write("FLAG_" + nombre_proc + " DW 0\n");
+                    myFile.write("nombre_" + nombre_proc + " DB " + "\'" + nombre_proc.substring(0, nombre_proc.indexOf("@")) + "\',0\n");
                 }
             }
         }
-        if(Salida.integer)
+        if (Salida.integer)
             myFile.write("aux_salida DD ?\n");
-        if(Comparador.comp_float || Division.hay_division_float){
+        if (Comparador.comp_float || Division.hay_division_float) {
             myFile.write("mem2bytes DW ?\n");
-            if(!lexico.tablaDeSimbolos.containsKey("0.0"))
+            if (!lexico.tablaDeSimbolos.containsKey("0.0"))
                 myFile.write("_0_0 DQ 0\n");
         }
-        if(Division.hay_division)
+        if (Division.hay_division)
             myFile.write("_cero DB 'Error: Division por cero',0\n");
-        if(Llamada.hay_llamada) {
+        if (Llamada.hay_llamada) {
             myFile.write("_recursion DB 'Error: Llamado recursivo de',0\n");
             myFile.write("_invocacion DB 'Error: Se alcanzo el numero maximo de invocaciones permitidas para el procedimiento',0\n");
             myFile.write("nombre_proc DD ?\n");
@@ -118,9 +125,9 @@ public class App {
         myFile.write("\n");
         myFile.write(".CODE\n");
         myFile.write(DecProc.procs.toString());
-        if(Division.hay_division)
+        if (Division.hay_division)
             myFile.write("_CERO:\ninvoke printf, cfm$(\"%s\\n\"),OFFSET _cero\nJMP _END\n");
-        if(Llamada.hay_llamada) {
+        if (Llamada.hay_llamada) {
             myFile.write("_RECURSION:\ninvoke printf, cfm$(\"%s %s\\n\"),OFFSET _recursion,nombre_proc\nJMP _END\n");
             myFile.write("_INVOCACION:\ninvoke printf, cfm$(\"%s %s\\n\"),OFFSET _invocacion,nombre_proc\nJMP _END\n");
         }
@@ -133,8 +140,8 @@ public class App {
 
         ProcessBuilder processBuilder = new ProcessBuilder();
         processBuilder.directory(new File(masm_path));
-        processBuilder.command("cmd.exe", "/c" ,
-                        "copy \"" + path + filename + ".asm\" \"" + filename + ".asm\" " +
+        processBuilder.command("cmd.exe", "/c",
+                "copy \"" + path + filename + ".asm\" \"" + filename + ".asm\" " +
                         "& ml /c /coff \"" + filename + ".asm\" " +
                         "& link /SUBSYSTEM:CONSOLE \"" + filename + ".obj\" " +
                         "& del /f \"" + filename + ".asm\" " +
